moves sensor lookup table to its own function, and also changes the size to hardcoded so no need to use pointer to array, or malloc
