+++ firmware/freedom/helpers/lookup_table.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "../freedom.h"
+
+void create_lookup_table(kb_config_t *kb_config, uint8_t sensor_lookup_table[SENSOR_COUNT][MAX_ADC_READING]);
