+++ firmware/freedom/rules.mk
@@ -14,3 +14,4 @@ RGBLIGHT_DRIVER = ws2812
 SRC += matrix.c
 SRC += analog.c
 SRC += helpers/sensor_read.c
+SRC += helpers/lookup_table.c
